[TVMC] Split common tvmc test file into more specific files
The `test_tvmc_common.py` file was becoming a bit of a mixed bag of
tests and as we now want to extend the `Target` processing logic it made
sense to split each out into its own file to make it clearer what each
does.

`test_common.py` has also been renamed before we start using it for all the
tests instead.
